+++ packages/javascript/lib/web/index.ts
@@ -1,5 +1,6 @@
 export { generateLicense } from './generateLicense.js';
 export { getParsedInfoFromAddress } from './getParsedInfoFromAddress.js';
+export { getSlug } from './getSlug.js';
 export { isBotAgent } from './isBotAgent.js';
 export { isMatchPathname } from './isMatchPathname.js';
 export { isMobile } from './isMobile.js';